Add some intial ideas and code for categories
This introduces the idea of 'category groups'. I am taking "TypeText"
and grouping them into categories that might be interesting to look at.

More work needs to be put into thinking about this. Here are a few
things off the top of my head:

1. Change the context from "crime" to "calls for service" since that is
what it is now. If we want to make this only about crime we will need
more filtering and also be cognizant of the fact that these are calls
and not necessarily "crimes".
2. Decide what data should be on the map for "All".
3. Decide which goes into what category group. `See categories.py`
4. Decide what categories would be interesting to people
5. Add visually tuned heatmap parameters for each category (for example, categories
with a small count should amplify the heatmap since right now you can't
see them well)